Combine actions and constants
