--- src/actions/callNotes.js
@@ -1,7 +1,0 @@
-import * as constants from 'src/constants/callNotes';
-
-
-export const createCallNotes = callId => ({
-  type: constants.CALL_NOTES_CREATE,
-  payload: { callId },
-});
