+++ src/constants/callNote.js
@@ -14,13 +14,14 @@ const CALL_NOTE_PATCH_REQUEST = 'CALL_NOTE_PATCH_REQUEST';
 const CALL_NOTE_PATCH_SUCCESS = 'CALL_NOTE_PATCH_SUCCESS';
 const CALL_NOTE_PATCH_FAILURE = 'CALL_NOTE_PATCH_FAILURE';
 
-
 const CALL_NOTES_MENTEE_HAPPY = 'happy';
 const CALL_NOTES_MENTEE_SAD = 'sad';
 const CALL_NOTES_MENTEE_BORED = 'bored';
 const CALL_NOTES_MENTEE_CONFUSED = 'confused';
 const CALL_NOTES_MENTEE_UPSET = 'upset';
 const CALL_NOTES_MENTEE_WITHDRAWN = 'withdrawn';
+
+const CALL_NOTES_CREATE = 'CALL_NOTES_CREATE';
 
 
 export {
@@ -46,4 +47,6 @@ export {
   CALL_NOTES_MENTEE_CONFUSED,
   CALL_NOTES_MENTEE_UPSET,
   CALL_NOTES_MENTEE_WITHDRAWN,
+
+  CALL_NOTES_CREATE,
 };
